backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
yfinance>=0.2.40
pandas==2.2.0
numpy==1.26.3
pytest==8.0.0
pytest-cov==4.1.0
pytest-asyncio==0.23.3
httpx==0.26.0
requests-mock==1.11.0
